Add files via upload
The project data focuses on the five social indicators named Population, Health, Housing, Education, and Work. The dataset contains multiple files for each indicator and for different countries. 
analyze how these factors are dependent on each other and make predictions based on the analysis performed. To begin with, we cleaned and prepared the data for the two indicators: health and work. 
Then the Decision Tree classifier is used based on Scikit Library on both the health and work data file. The dependent Variable is classified based on global mean into two categories.